feat: #77 BoostAnimationType 분리
